+++ src/kkbding2-bundle/Resources/public/apps/dingEditors/ding-events-editor.js
@@ -1,0 +1,12 @@
+  angular.module('toolsModule').directive('dingEventsEditor', function(){
+  return {
+    restrict: 'E',
+    replace: true,
+    scope: {
+      slide:'=',
+      close: '&',
+      template: '@'
+    },
+    templateUrl: '/bundles/kkbding2integration/apps/dingEditors/ding-events-editor.html'
+  };
+});