+++ src/kkbding2-bundle/Resources/public/apps/dingEditors/info-editor.js
@@ -1,0 +1,12 @@
+angular.module('toolsModule').directive('infoEditor', function(){
+  return {
+    restrict: 'E',
+    replace: true,
+    scope: {
+      slide:'=',
+      close: '&',
+      template: '@'
+    },
+    templateUrl: '/bundles/kkbding2integration/apps/dingEditors/info-editor.html'
+  };
+});