+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # CHANGELOG
 
+## 6.1.0
+
+* Added screen-bundle.
+
+## 6.0.0
+
+* Upgraded to symfony 3.4.
+
+## 5.2.0
+
+* Updated symfony/symfony: v2.8.49
+* Updated os2display/core-bundle: 1.2.0
+* Updated os2display/admin-bundle: 1.2.0
+* Updated os2display/default-template-bundle: 1.1.1
+
+## 5.1.2
+
+* Merged https://github.com/os2display/admin/pull/17
+
+## 5.1.1
+
+* Updated symfony/symfony: 2.8.44
+* Updated os2display/admin-bundle: 1.1.1
+
 ## 5.1.0
 
 * Updated symfony/symfony: 2.8.42